requirements.txt
python-telegram-bot==20.7
gspread==5.8.0
oauth2client==4.1.3
google-auth==2.6.0
